Creates the logic for sending courses in addition to projects.

+++ src/controllers/OrderController.js
@@ -52,7 +52,21 @@ module.exports = {
     const db = admin.firestore();
     const storage = admin.storage().bucket();
 
-    let projectsToSend = await Promise.all(order.itens.map(async (item) => {
+    const coursesToSend = (await Promise.all(order.itens.map(async (item) => {
+      const courseData = await db.collection('courses').doc(item.sku).get();
+      const course = courseData.data();
+
+      if (!course) {
+        return false;
+      }
+
+      return {
+        sku: item.sku,
+        ...course,
+      };
+    }))).filter((item) => item);
+
+    const projectsToSend = (await Promise.all(order.itens.filter((item) => !coursesToSend.map((course) => course.sku).includes(item.sku)).map(async (item) => {
       const projectData = await db.collection('projects').doc(item.sku).get();
       const project = projectData.data();
 
@@ -89,14 +103,12 @@ module.exports = {
       }));
 
       return [...regularFiles, ...encryptableFiles];
-    }));
+    }))).filter((item) => item);
 
-    projectsToSend = projectsToSend.filter((item) => item);
-
-    if (projectsToSend.length === 0) {
+    if (projectsToSend.length === 0 && coursesToSend.length === 0) {
       fs.rmdirSync(directory, { recursive: true });
       return response.status(400).json({
-        error: 'NO_PROJECTS_TO_SEND',
+        error: 'NO_PRODUCTS_TO_SEND',
         status: null,
       });
     }
@@ -120,22 +132,55 @@ module.exports = {
           tls: { rejectUnauthorized: false },
         });
 
-        const html = `
-        <p>Olá ${order.cliente.nome}! Tudo bem?</p>
-        <p>Aqui estão os arquivos dos projetos que você comprou na nossa loja.</p>
-        <br>
-        <p>Os PDFs são salvos com uma senha única, que é seu CPF/CNPJ sem pontos ou traços, conforme abaixo:</p>
-        <p><strong>${password}</strong></p>
-        <br>
-        <p>Para qualquer dúvida ou necessidade, entre em contato conosco através deste e-mail ou via WhatsApp no número +55 13 98219-9905</p>
-        <br>
-        <p>Obrigado pela confiança!</p>
-        <p>Equipe Artes no Reino</p>
-        <br>
-        <a href="https://loja.artesnoreino.com.br">Clique aqui para acessar a nossa lojinha</a>`;
+        let html = `
+          <p>Olá ${order.cliente.nome.trim()}! Tudo bem?</p>
+        `;
+
+        if (coursesToSend.length === 0) {
+          html += `
+            <p>Aqui estão os arquivos dos projetos que você comprou na nossa loja.</p>
+            <br>
+            <p>Os PDFs são salvos com uma senha única, que é seu CPF/CNPJ sem pontos ou traços, conforme abaixo:</p>
+            <p><strong>${password}</strong></p>
+            <br>
+            <p>Para qualquer dúvida ou necessidade, entre em contato conosco através deste e-mail ou via WhatsApp no número +55 13 98219-9905</p>
+            <br>`;
+        } else if (projectsToSend.length === 0) {
+          html += `
+            <p>Aqui estão os links dos cursos que você comprou na nossa loja.</p>
+            <p>Para se inscrever basta acessar cada um destes links através do Facebook e enviar a solicitação de entrada no grupo, e assim que possível nossa equipe irá aceitar o seu pedido.</p>`;
+
+          coursesToSend.forEach((item) => {
+            html += `<p><strong>${item.name}</strong>: ${item.link}</p>`;
+          });
+
+          html += '<br>';
+        } else {
+          html += `
+            <p>Aqui estão os arquivos dos projetos que você comprou na nossa loja.</p>
+            <br>
+            <p>Os PDFs são salvos com uma senha única, que é seu CPF/CNPJ sem pontos ou traços, conforme abaixo:</p>
+            <p><strong>${password}</strong></p>
+            <br>
+            <p>Aqui também estão os links dos cursos que você comprou na nossa loja.</p>
+            <p>Para se inscrever basta acessar cada um destes links através do Facebook e enviar a solicitação de entrada no grupo, e assim que possível nossa equipe irá aceitar o seu pedido.</p>`;
+
+          coursesToSend.forEach((item) => {
+            html += `<p><strong>${item.name}</strong>: ${item.link}</p>`;
+          });
+        }
+
+        html += `<br>
+          <p>Para qualquer dúvida ou necessidade, entre em contato conosco através deste e-mail ou <a href="https://api.whatsapp.com/send?phone=5513982199905&text=Ol%C3%A1%2C%20sou%20${order.cliente.nome.trim()}%20e%20gostaria%20de%20falar%20sobre%20o%20pedido%20${id}.">via WhatsApp no número +55 13 98219-9905</a></p>
+          <br>
+          <p>Obrigado pela confiança!</p>
+          <p>Equipe Artes no Reino</p>
+          <br>
+          <a href="https://loja.artesnoreino.com.br">Clique aqui para acessar a nossa lojinha</a>
+        `;
 
         const mailOptions = {
-          from: `"Artes no Reino" <${process.env.GMAIL_USER}`,
+          from: `"Artes no Reino" <${process.env.GMAIL_USER}>`,
           to: order.cliente.email,
           subject: `Pedido de venda #${id}`,
           html,
@@ -159,7 +204,7 @@ module.exports = {
       });
     }
 
-    const codigo = order.itens.length !== projectsToSend.length ? 'em_producao' : 'pedido_entregue';
+    const codigo = order.itens.length !== [...projectsToSend, ...coursesToSend].length ? 'em_producao' : 'pedido_entregue';
 
     await axios.put(`situacao/pedido/${id}`, {
       codigo,
